Rename gzip module to compression witout config changes
